# src/integration-test/resources/openapi.yml
openapi: 3.1.0
info:
  title: api-comfact-facade
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "3.0"
servers:
  - url: http://localhost:58747
    description: Generated server url
tags:
  - name: Signings
    description: Signing operations
paths:
  /{municipalityId}/signings:
    get:
      tags:
        - Signings
      summary: Get all signing instances.
      description: "The 'sort' parameter in the Pageable object can take the values\
        \ 'SigningInstanceId', 'ReferenceNumber', 'CustomerReferenceNumber', 'Created',\
        \ 'Changed', 'Expires', 'StatusCode', 'StatusMessage', 'UserId', 'Language',\
        \ 'SignatoryReminderStartDate', 'InitiatorEmailAddress', 'QueueCreated', 'QueueChanged'.\
        \ Will default to 'SigningInstanceId' if not set. Will default to page 0 with\
        \ a size of 10 if not set."
      operationId: getSigningRequests
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigningsResponse"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - Signings
      summary: Create Signing instance.
      operationId: createSigningRequest
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigningRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSigningResponse"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/signings/{signingId}:
    get:
      tags:
        - Signings
      summary: Get a signing request.
      operationId: getSigningRequest
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: signingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigningInstance"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
        - Signings
      summary: Update a signing instance.
      operationId: updateSigningRequest
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: signingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSigningRequest"
        required: true
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/signings/{signingId}/signatory/{partyId}:
    get:
      tags:
        - Signings
      summary: Get information about the current signatory
      operationId: getSignatory
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: signingId
          in: path
          required: true
          schema:
            type: string
        - name: partyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signatory"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api-docs:
    get:
      tags:
        - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Document:
      type: object
      properties:
        name:
          type: string
          description: The Document Name visible to the user in the signing process.
          examples:
            - Business Contract
        fileName:
          type: string
          description: "The document file name including extension,"
          examples:
            - document.pdf
          minLength: 1
        mimeType:
          type: string
          description: The documents mimetype. Must be application/pdf
          examples:
            - application/pdf
        content:
          type: string
          format: byte
          description: Base64-encoded file (plain text)
          examples:
            - dGVzdA==
      required:
        - fileName
    Identification:
      type: object
      description: The means of identification to use to identify the signatories.
      properties:
        alias:
          type: string
          description: |
            Possible values for Alias type:
            • SmsCode – One Time Password via SMS
            • EmailCode – One Time Password via Email
            • SvensktEId – Swedish e-identification (Swedish BankID)
          examples:
            - SmsCode
    NotificationMessage:
      type: object
      description: Custom message for the signature request emails.
      properties:
        subject:
          type: string
          description: The subject of the notification message
          examples:
            - Please sign the document
        body:
          type: string
          description: The body of the notification message
          examples:
            - "Dear John Doe, please sign the document."
        language:
          type: string
          description: The language of the notification message.
          examples:
            - sv-SE
          minLength: 1
      required:
        - language
    Party:
      type: object
      description: A party related to the signing process.
      properties:
        name:
          type: string
          description: The name of the party
          examples:
            - John Doe
        partyId:
          type: string
          description: The party id
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
        notificationMessage:
          $ref: "#/components/schemas/NotificationMessage"
          description: Custom message for the signature request emails for the specific
            party. Overwrites the message in the Signing Request.
        title:
          type: string
          description: The party title
          examples:
            - CEO
        email:
          type: string
          description: The party email
          examples:
            - john.doe@sundsvall.se
        phoneNumber:
          type: string
          description: The party phone number
          examples:
            - "0701740605"
        organization:
          type: string
          description: The organization of the party.
          examples:
            - Sundsvall Municipality
        language:
          type: string
          description: Language parameter that overwrites the language of the Signing
            Instance for the current party.
          examples:
            - sv
      required:
        - email
    Reminder:
      type: object
      description: A reminder for a signature request.
      properties:
        message:
          $ref: "#/components/schemas/NotificationMessage"
          description: The message of the reminder
        enabled:
          type: boolean
          description: If the reminder is enabled
          examples:
            - "true"
        intervalInHours:
          type: integer
          format: int32
          description: The interval in hours between each reminder
          examples:
            - "24"
        startDateTime:
          type: string
          format: date-time
          description: The date and time when the first reminder should be sent.
          examples:
            - 2026-11-22T15:30:00+02:00
      required:
        - enabled
        - intervalInHours
    Signatory:
      type: object
      description: A Signatory related to the signing process.
      properties:
        name:
          type: string
          description: The name of the party
          examples:
            - John Doe
        partyId:
          type: string
          description: The party id
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
        notificationMessage:
          $ref: "#/components/schemas/NotificationMessage"
          description: Custom message for the signature request emails for the specific
            party. Overwrites the message in the Signing Request.
        title:
          type: string
          description: The party title
          examples:
            - CEO
        email:
          type: string
          description: The party email
          examples:
            - john.doe@sundsvall.se
        phoneNumber:
          type: string
          description: The party phone number
          examples:
            - "0701740605"
        organization:
          type: string
          description: The organization of the party.
          examples:
            - Sundsvall Municipality
        language:
          type: string
          description: Language parameter that overwrites the language of the Signing
            Instance for the current party.
          examples:
            - sv
        identifications:
          type: array
          items:
            $ref: "#/components/schemas/Identification"
          minItems: 1
      required:
        - email
        - identifications
    SigningRequest:
      type: object
      description: The signing request.
      properties:
        additionalDocuments:
          type: array
          items:
            $ref: "#/components/schemas/Document"
            description: Additional documents to sign.
        language:
          type: string
          description: The language of the signing process.
          examples:
            - sv-SE
        customerReference:
          type: string
          description: The customer reference
          examples:
            - "1234567890"
        expires:
          type: string
          format: date-time
          description: The date and time when the signing request expires.
          examples:
            - 2026-11-22T15:30:00+02:00
        notificationMessage:
          $ref: "#/components/schemas/NotificationMessage"
          description: Custom message for the signature request emails.
        reminder:
          $ref: "#/components/schemas/Reminder"
          description: A reminder for a signature request.
        initiator:
          $ref: "#/components/schemas/Party"
          description: The initiator of the signing request.
        additionalParties:
          type: array
          description: A party that is not part of the signing process but will get
            a copy of the signed document when the signing instance is completed.
          items:
            $ref: "#/components/schemas/Party"
        signatories:
          type: array
          description: The parties that will sign the documents.
          items:
            $ref: "#/components/schemas/Signatory"
          minItems: 1
        document:
          $ref: "#/components/schemas/Document"
          description: Information about the main document to sign.
        flowType:
          type: string
          description: The flow type of the signing process.
          examples:
            - Sequential
      required:
        - document
        - initiator
        - signatories
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    CreateSigningResponse:
      type: object
      properties:
        signingId:
          type: string
        signatoryUrls:
          type: object
          additionalProperties:
            type: string
    UpdateSigningRequest:
      type: object
      description: Request for updating a signing instance.
      properties:
        expires:
          type: string
          format: date-time
          description: The date and time when the signing request expires.
          examples:
            - 2026-11-22T15:30:00+02:00
        status:
          type: string
          description: The status of the signing instance.
          examples:
            - active
            - withdrawn
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    Direction:
      type: string
      enum:
        - ASC
        - DESC
    PagingAndSortingMetaData:
      type: object
      description: PagingAndSortingMetaData model
      properties:
        page:
          type: integer
          format: int32
          description: Current page
          examples:
            - "5"
          readOnly: true
        limit:
          type: integer
          format: int32
          description: Displayed objects per page
          examples:
            - "20"
          readOnly: true
        count:
          type: integer
          format: int32
          description: Displayed objects on current page
          examples:
            - "13"
          readOnly: true
        totalRecords:
          type: integer
          format: int64
          description: Total amount of hits based on provided search parameters
          examples:
            - "98"
          readOnly: true
        totalPages:
          type: integer
          format: int32
          description: Total amount of pages based on provided search parameters
          examples:
            - "23"
          readOnly: true
        sortBy:
          type: array
          items:
            type: string
            description: The properties to sort by
            examples:
              - property
            readOnly: true
        sortDirection:
          $ref: "#/components/schemas/Direction"
          description: The sort order direction
          examples:
            - ASC
    SigningInstance:
      type: object
      description: The signing instance.
      properties:
        notificationMessages:
          type: array
          description: Custom message for the signature request emails.
          items:
            $ref: "#/components/schemas/NotificationMessage"
        initiator:
          $ref: "#/components/schemas/Party"
          description: The initiator of the signing request.
        additionalParties:
          type: array
          description: A party that is not part of the signing process but will get
            a copy of the signed document when the signing instance is completed.
          items:
            $ref: "#/components/schemas/Party"
        signatories:
          type: array
          description: The parties that will sign the documents.
          items:
            $ref: "#/components/schemas/Signatory"
          minItems: 1
        status:
          $ref: "#/components/schemas/Status"
          description: Status of the signing request.
        customerReference:
          type: string
          description: The customer reference
          examples:
            - "1234567890"
        created:
          type: string
          format: date-time
          description: The date and time when the signing instance was created.
          examples:
            - 2026-04-02T11:20:00+02:00
        changed:
          type: string
          format: date-time
          description: The date and time when the signing instance was last changed.
          examples:
            - 2026-04-02T11:20:00+02:00
        expires:
          type: string
          format: date-time
          description: The date and time when the signing instance expires
          examples:
            - 2026-11-22T15:30:00+02:00
        document:
          $ref: "#/components/schemas/Document"
          description: Information about the main document to sign.
        signedDocument:
          $ref: "#/components/schemas/Document"
          description: Information about the signed document
        additionalDocuments:
          type: array
          items:
            $ref: "#/components/schemas/Document"
            description: Additional documents to sign.
        signingId:
          type: string
          description: The signing instance id.
          examples:
            - "1234567890"
      required:
        - initiator
        - signatories
    SigningsResponse:
      type: object
      description: Signing response.
      properties:
        signingInstances:
          type: array
          description: The signing instances.
          items:
            $ref: "#/components/schemas/SigningInstance"
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
    Status:
      type: object
      properties:
        code:
          type: string
          description: The status code
          examples:
            - Created
        message:
          type: string
          description: The status message
          examples:
            - The signing instance is halted.
  securitySchemes: {}